scan_me_right/ml_training/requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
tensorflow>=2.13.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
matplotlib>=3.7.0

